yourfilename.py
import logging
import asyncio
from telegram import Update, InlineKeyboardButton, InlineKeyboardMarkup
from telegram.ext import (
    ApplicationBuilder,
    CommandHandler,
    CallbackQueryHandler,
    ContextTypes
)

# Set up logging
logging.basicConfig(format='%(asctime)s - %(levelname)s - %(message)s',
                    level=logging.INFO)
logger = logging.getLogger(__name__)

# === Command Handlers ===

# Start Command
async def start(update: Update, context: ContextTypes.DEFAULT_TYPE):
    await update.message.reply_text("Welcome to the Forex Bot!")

# Status Command
async def status(update: Update, context: ContextTypes.DEFAULT_TYPE):
    await update.message.reply_text("Bot is up and running!")

# Analyze Command
async def analyze(update: Update, context: ContextTypes.DEFAULT_TYPE):
    symbol = context.args[0] if context.args else "EURUSD"
    # Perform analysis here and send the result
    analysis_result = f"Analysis for {symbol}: Buy Signal"
    await update.message.reply_text(analysis_result)

# Summary Command
async def summary(update: Update, context: ContextTypes.DEFAULT_TYPE):
    try:
        # Generate summary (replace with your logic)
        report = "Forex market summary for the day."
        await update.message.reply_text(report)
    except Exception as e:
        logger.exception("Summary failed")
        await update.message.reply_text("⚠️ Summary failed")

# Chart Command
async def chart(update: Update, context: ContextTypes.DEFAULT_TYPE):
    # Placeholder for chart generation logic
    chart_url = "https://www.example.com/chart"  # Replace with actual chart URL generation
    await update.message.reply_text(f"Chart: {chart_url}")

# News Command
async def news(update: Update, context: ContextTypes.DEFAULT_TYPE):
    # Placeholder for news fetching
    news_update = "Latest Forex news update."
    await update.message.reply_text(news_update)

# Update Command
async def update_now(update: Update, context: ContextTypes.DEFAULT_TYPE):
    # Placeholder for update logic
    await update.message.reply_text("Refreshing data cache...")

# === Callback Handlers ===
async def handle_callback(update: Update, context: ContextTypes.DEFAULT_TYPE):
    query = update.callback_query
    cmd = query.data
    context.args = [cmd.split('_')[1]]
    await analyze(update, context)

# === Main Entry ===
def main():
    app = ApplicationBuilder().token('YOUR_BOT_TOKEN').build()

    app.add_handler(CommandHandler("start", start))
    app.add_handler(CommandHandler("status", status))
    app.add_handler(CommandHandler("analyze", analyze))
    app.add_handler(CommandHandler("summary", summary))
    app.add_handler(CommandHandler("chart", chart))
    app.add_handler(CommandHandler("news", news))
    app.add_handler(CommandHandler("update", update_now))
    app.add_handler(CallbackQueryHandler(handle_callback))

    app.run_polling()

if __name__ == "__main__":
    try:
        main()
    except Exception as e:
        logger.exception("Fatal bot error")
