requirements.txt
# Termux-stable pins (avoid curl_cffi/libcurl-impersonate issues)
yfinance==0.2.40
pandas==1.5.3
numpy==1.26.4
python-dateutil>=2.8.2
requests>=2.31.0

# your bot deps
python-telegram-bot==20.7
